Move ReverseMapper logic onto OpaqueHiddenType
